pyproject.toml
[tool.poetry]
name = "django-fs"
version = "0.1.0"
description = ""
authors = ["vlf0 <dr.dnb2000@gmail.com>"]
readme = "README.md"

[tool.pytest.ini_options]
addopts = "-q --cov"
DJANGO_SETTINGS_MODULE = "django_fs.settings"
python_files = "test.py tests.py"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1.2"
djangorestframework = "3.15.2"
python-dotenv = "1.0.1"
pytest = "8.3.3"
ruff = "0.6.9"
django-environ = "0.11.2"
psycopg2-binary = "2.9.9"
pytest-django = "4.9.0"
pytest-cov = "5.0.0"
drf-spectacular = "0.27.2"

[tool.coverage.run]
branch = true
source = ["transactions", "users"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"